push presensi page bisa edit now

+++ lib/data/userHistory.dart
@@ -13,14 +13,13 @@ class UserHistory {
   final String? tipe_delivery;
   final int? penggunaan_poin;
   final int? total;
-  final double? radius;
+  final int? radius;
   final int? ongkir;
   final String? status;
   final int? penambahan_poin;
   final int? poin_user_setelah_penambahan;
-  final List<Produk>? detailTransaksi; 
-  final List<Produk>? detailTransaksi1; 
-  
+  final List<Produk>? detailTransaksi;
+  final List<Produk>? detailTransaksi1;
 
   UserHistory({
     this.no_nota,
@@ -40,8 +39,7 @@ class UserHistory {
     this.status,
     this.penambahan_poin,
     this.poin_user_setelah_penambahan,
-    
-    this.detailTransaksi, 
+    this.detailTransaksi,
     this.detailTransaksi1,
   });
 
@@ -66,14 +64,19 @@ class UserHistory {
       status: json['status'],
       penambahan_poin: json['penambahan_poin'],
       poin_user_setelah_penambahan: json['poin_user_setelah_penambahan'],
-      detailTransaksi: json['produk'] != null ? List<Produk>.from(json['produk'].map((x) => Produk.fromJson(x))) : null, 
-      detailTransaksi1: json['detail_transaksi'] != null ? List<Produk>.from(json['detail_transaksi'].map((x) => Produk.fromJson(x))) : null, 
+      detailTransaksi: json['produk'] != null
+          ? List<Produk>.from(json['produk'].map((x) => Produk.fromJson(x)))
+          : null,
+      detailTransaksi1: json['detail_transaksi'] != null
+          ? List<Produk>.from(
+              json['detail_transaksi'].map((x) => Produk.fromJson(x)))
+          : null,
     );
   }
 
   Map<String, dynamic> toJson() => {
         "no_nota": no_nota,
-        "nama" : nama,
+        "nama": nama,
       };
 }
 
@@ -102,8 +105,3 @@ class Produk {
     );
   }
 }
-
-
-
-
-
